refactors concern's polymorphic helpers
